+++ test/mock/chimerax/core/commands/__init__.py
@@ -1,0 +1,15 @@
+class CmdDesc(object):
+    def __init__(self, required=(), optional=(), keyword=(),
+                 postconditions=(), required_arguments=(),
+                 non_keyword=(), hidden=(), url=None, synopsis=None):
+        self.synopsis = synopsis
+
+class IntArg:
+    pass
+
+class ModelArg:
+    pass
+
+
+def register(name, cmd_desc=(), function=None, *, logger=None, registry=None):
+    pass