+++ ChangeLog.rst
@@ -2,6 +2,7 @@ HEAD
 ====
  - Improved support for RMF files containing atomic information.
  - Add basic visualization of RMF geometry, such as bounding boxes.
+ - Add an `rmf hierarchy` command to show the RMF hierarchy in the log.
 
 0.1 - 2020-02-19
 ================